feat: when choose film from autocomplete, markers are set
This took hours to figure out. Basically, don't use data-bind = textInput: singleFilm because that will give you the input, such as when you click 'A' it will store 'A' when what you really want is the ultimate value after you've chosen from a series of films that all start with 'A'. So, use data-bind="value: singleFilm" because it gives you that ultimate value, the full movie name. Then I added the button back. I don't have to use it, I can use something else and probably will. Either subscribe like this:
// my.vm.singleFilm.subscribe(function() {
//   my.vm.codeAddress();
// }, my.vm);

or I might go back to the bootstrap 'go' button.